backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pymupdf
python-docx==1.1.0
python-pptx==0.6.23
Pillow
opencv-python-headless
numpy
aiofiles==23.2.1
